requirements.txt
aws-cdk-lib==2.27.0
constructs>=10.0.0,<11.0.0
pydantic==1.9.1
